+++ src/SpamOK.PasswordGenerator/Models/PasswordEntropy.cs
@@ -58,7 +58,7 @@ namespace SpamOK.PasswordGenerator.Models
         /// Get average time to crack the password in seconds.
         /// </summary>
         /// <returns>Time to crack password in seconds as integer.</returns>
-        public int GetTimeToCrackSeconds()
+        public long GetTimeToCrackSeconds()
         {
             return EntropyCalculatorHelper.GetTimeToCrack(BitEntropy);
         }
